--- extensions/crawler/package.json
@@ -1,39 +1,0 @@
-{
-	"name": "crawler",
-	"description": "Please enter a description for your extension",
-	"icon": "extension",
-	"version": "1.0.0",
-	"keywords": [
-		"directus",
-		"directus-extension",
-		"directus-extension-hook"
-	],
-	"type": "commonjs",
-	"files": [
-		"dist"
-	],
-	"directus:extension": {
-		"type": "hook",
-		"path": "dist/index.js",
-		"source": "src/index.ts",
-		"host": "^10.10.0"
-	},
-	"scripts": {
-		"build": "directus-extension build",
-		"dev": "directus-extension build -w --no-minify",
-		"link": "directus-extension link",
-		"validate": "directus-extension validate"
-	},
-	"devDependencies": {
-		"@directus/api": "^25.0.1",
-		"@directus/extensions-sdk": "13.0.3",
-		"@directus/sdk": "^19.0.1",
-		"@directus/types": "^13.0.0",
-		"@types/node": "^22.13.10",
-		"typescript": "^5.8.2"
-	},
-	"dependencies": {
-		"csvjson": "^5.1.0",
-		"moment": "^2.30.1"
-	}
-}
