--- extensions/crawler/src/types/ICrawler.ts
@@ -1,4 +1,0 @@
-export default interface ICrawler {
-  // entry point for a crawler to run
-  crawl: () => void
-}
