--- extensions/crawler/src/types/ICsvParameters.ts
@@ -1,3 +1,0 @@
-export default interface ICsvParameters {
-  delimiter: string
-}
