+++ extensions/o-mate/src/race-crawler/app.ts
@@ -1,0 +1,15 @@
+import { defineOperationApp } from '@directus/extensions-sdk'
+
+export default defineOperationApp({
+  id: 'race-crawler',
+  name: 'Race Crawler',
+  icon: 'box',
+  description: 'Crawl for orienteering races.',
+  overview: ({ text }) => [
+    {
+      label: 'Text',
+      text: text,
+    },
+  ],
+  options: [],
+})
