+++ extensions/o-mate/src/race-crawler/shims.d.ts
@@ -1,0 +1,5 @@
+declare module '*.vue' {
+	import { DefineComponent } from 'vue';
+	const component: DefineComponent<{}, {}, any>;
+	export default component;
+}
