+++ extensions/o-mate/src/race-crawler/types/ICrawler.ts
@@ -1,0 +1,4 @@
+export default interface ICrawler {
+  // entry point for a crawler to run
+  crawl: () => void
+}
