+++ extensions/o-mate/src/race-crawler/types/ICsvParameters.ts
@@ -1,0 +1,3 @@
+export default interface ICsvParameters {
+  delimiter: string
+}
