+++ extensions/o-mate/src/types/Utilities.ts
@@ -1,0 +1,7 @@
+import { Schema } from './DirectusTypes'
+
+export type PropertiesOnly<T> = {
+  [K in keyof T as T[K] extends Function ? never : K]: T[K]
+}
+
+export type TableName = keyof Schema
